backend/backend-requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
coverage==6.3.2
debugpy==1.5.1
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-rest-registration==0.7.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
Faker==13.3.0
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython==8.1.1
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
regex==2022.3.2
scikit-surprise==1.1.1
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
surprise==0.1
text-unidecode==1.3
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing-extensions==4.1.1
wcwidth==0.2.5
zipp==3.7.0
